refactor the logic of fetching url and blobs
