+++ frontend/src/components/visualizations/Audio/Audio.tsx
@@ -1,15 +1,9 @@
 import React, { useEffect, useState } from "react";
-import { useDispatch, useSelector } from "react-redux";
-import { RootState } from "../../../types/data";
+import { useDispatch } from "react-redux";
 import {
-	GENERATE_FILE_URL,
-	generateFileDownloadUrl as generateFileDownloadUrlAction,
-} from "../../../actions/file";
-
-import {
-	GENERATE_VIS_URL,
-	generateVisDataDownloadUrl as generateVisDataDownloadUrlAction,
-} from "../../../actions/visualization";
+	generateFileDownloadUrl,
+	generateVisDataDownloadUrl,
+} from "../../../utils/visualization";
 
 type AudioProps = {
 	fileId?: string;
@@ -21,44 +15,25 @@ export default function Audio(props: AudioProps) {
 
 	const dispatch = useDispatch();
 
-	const generateFileDownloadUrl = (
-		fileId: string | undefined,
-		fileVersionNum: number | undefined
-	) => dispatch(generateFileDownloadUrlAction(fileId, fileVersionNum));
-
-	const generateVisDataDownloadUrl = (visualizationId: string | undefined) =>
-		dispatch(generateVisDataDownloadUrlAction(visualizationId));
-
-	const rawFileURL = useSelector((state: RootState) => state.file.url);
-	const visFileURL = useSelector((state: RootState) => state.visualization.url);
-
 	const [url, setUrl] = useState("");
 
 	useEffect(() => {
-		// reset
-		return () => {
-			dispatch({
-				type: GENERATE_FILE_URL,
-				url: "",
-				receivedAt: Date.now(),
-			});
-			dispatch({
-				type: GENERATE_VIS_URL,
-				url: "",
-				receivedAt: Date.now(),
-			});
+		const fetchUrl = async () => {
+			try {
+				let downloadUrl;
+				if (visualizationId) {
+					downloadUrl = await generateVisDataDownloadUrl(visualizationId);
+				} else {
+					downloadUrl = await generateFileDownloadUrl(fileId, 0);
+				}
+				setUrl(downloadUrl);
+			} catch (error) {
+				console.error("Error fetching data:", error);
+			}
 		};
-	}, []);
 
-	useEffect(() => {
-		if (visualizationId) generateVisDataDownloadUrl(visualizationId);
-		else generateFileDownloadUrl(fileId, 0);
+		fetchUrl();
 	}, [visualizationId, fileId]);
-
-	useEffect(() => {
-		if (visFileURL && visFileURL !== "") setUrl(visFileURL);
-		else setUrl(rawFileURL);
-	}, [visFileURL, rawFileURL]);
 
 	return (() => {
 		if (url && url !== "") {
